Updated MapperTests to test new tcp contracts and fixed regression in PartitionResponse.cs

+++ Iggy_SDK/Contracts/Http/PartitionResponse.cs
@@ -7,7 +7,7 @@ public sealed class PartitionContract
 
 	public required int SegmentsCount { get; init; }
 
-	public required int CurrentOffset { get; init; }
+	public required ulong CurrentOffset { get; init; }
 
-	public required int SizeBytes { get; init; }
+	public required ulong SizeBytes { get; init; }
 }
